+++ distant-local/src/api.rs
@@ -3,10 +3,10 @@ use std::time::SystemTime;
 use std::{env, io};
 
 use async_trait::async_trait;
+use distant_core::protocol::semver;
 use distant_core::protocol::{
-    Capabilities, ChangeKind, ChangeKindSet, DirEntry, Environment, FileType, Metadata,
-    Permissions, ProcessId, PtySize, SearchId, SearchQuery, SetPermissionsOptions, SystemInfo,
-    Version, PROTOCOL_VERSION,
+    ChangeKind, ChangeKindSet, DirEntry, Environment, FileType, Metadata, Permissions, ProcessId,
+    PtySize, SearchId, SearchQuery, SetPermissionsOptions, SystemInfo, Version, PROTOCOL_VERSION,
 };
 use distant_core::{DistantApi, DistantCtx};
 use ignore::{DirEntry as WalkDirEntry, WalkBuilder};
@@ -635,10 +635,32 @@ impl DistantApi for Api {
     async fn version(&self, ctx: DistantCtx) -> io::Result<Version> {
         debug!("[Conn {}] Querying version", ctx.connection_id);
 
+        // Parse our server's version
+        let mut server_version: semver::Version = env!("CARGO_PKG_VERSION")
+            .parse()
+            .map_err(|x| io::Error::new(io::ErrorKind::Other, x))?;
+
+        // Add the package name to the version information
+        if server_version.build.is_empty() {
+            server_version.build = semver::BuildMetadata::new(env!("CARGO_PKG_NAME"))
+                .map_err(|x| io::Error::new(io::ErrorKind::Other, x))?;
+        } else {
+            let raw_build_str = format!(
+                "{}.{}",
+                server_version.build.as_str(),
+                env!("CARGO_PKG_NAME")
+            );
+            server_version.build = semver::BuildMetadata::new(&raw_build_str)
+                .map_err(|x| io::Error::new(io::ErrorKind::Other, x))?;
+        }
+
         Ok(Version {
-            server_version: format!("{} {}", env!("CARGO_PKG_NAME"), env!("CARGO_PKG_VERSION")),
+            server_version,
             protocol_version: PROTOCOL_VERSION,
-            capabilities: Capabilities::all(),
+            capabilities: Version::capabilities()
+                .iter()
+                .map(ToString::to_string)
+                .collect(),
         })
     }
 }